implement get artist, reorg
