+++ spotify/client.go
@@ -77,40 +77,6 @@ func checkCache(ctx context.Context, key string) (*[]byte, error) {
 	return &bytes, nil
 }
 
-func getArtists(ctx context.Context, ids []string) (*Artists, error) {
-	token := ctx.Value(ContextAccessToken)
-	if token == nil {
-		return nil, errors.New("no access token provided")
-	}
-
-	url := fmt.Sprint("https://api.spotify.com/v1/artists?ids=", strings.Join(ids, ","))
-
-	req, err := http.NewRequest("GET", url, nil)
-	if err != nil {
-		return nil, err
-	}
-
-	req.Header.Add("Authorization", fmt.Sprint("Bearer ", token))
-
-	body, err := makeRequest(ctx, req, true)
-	if err != nil {
-		return nil, err
-	}
-
-	type tempResp struct {
-		Items Artists `json:"artists"`
-	}
-
-	var ret tempResp
-
-	err = json.Unmarshal(*body, &ret)
-	if err != nil {
-		return nil, err
-	}
-
-	return &ret.Items, nil
-}
-
 func getAudioFeatures(ctx context.Context, ids []string) (*AudioFeatures, error) {
 	logger := logging.GetLogger(&ctx)
 	token := ctx.Value(ContextAccessToken)
